Removing MagicLinkSignInActivity and consolidating all logic to SignInActivity
