+++ WordPress/src/main/java/org/wordpress/android/ui/accounts/SignInActivity.java
@@ -1,9 +1,13 @@
 package org.wordpress.android.ui.accounts;
 
+import android.app.ProgressDialog;
+import android.content.DialogInterface;
 import android.content.Intent;
 import android.net.Uri;
 import android.os.Bundle;
 import android.support.annotation.NonNull;
+import android.support.v4.app.Fragment;
+import android.support.v4.app.FragmentManager;
 import android.support.v4.app.FragmentTransaction;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Window;
@@ -17,13 +21,19 @@ import org.wordpress.android.R;
 import org.wordpress.android.WordPress;
 import org.wordpress.android.analytics.AnalyticsTracker;
 import org.wordpress.android.analytics.AnalyticsTracker.Stat;
+import org.wordpress.android.models.Account;
+import org.wordpress.android.models.AccountHelper;
 import org.wordpress.android.models.Blog;
 import org.wordpress.android.ui.ActivityId;
 import org.wordpress.android.ui.accounts.SmartLockHelper.Callback;
+import org.wordpress.android.ui.accounts.login.MagicLinkRequestFragment;
+import org.wordpress.android.ui.accounts.login.MagicLinkSentFragment;
+import org.wordpress.android.ui.accounts.login.MagicLinkSignInFragment;
 import org.wordpress.android.util.AppLog;
 import org.wordpress.android.util.AppLog.T;
 
-public class SignInActivity extends AppCompatActivity implements ConnectionCallbacks, OnConnectionFailedListener {
+public class SignInActivity extends AppCompatActivity implements ConnectionCallbacks, OnConnectionFailedListener, MagicLinkRequestFragment.OnMagicLinkFragmentInteraction,
+        MagicLinkSignInFragment.OnMagicLinkRequestInteraction, MagicLinkSentFragment.OnMagicLinkSentInteraction {
     public static final int SIGN_IN_REQUEST = 1;
     public static final int REQUEST_CODE = 5000;
     public static final int ADD_SELF_HOSTED_BLOG = 2;
@@ -36,8 +46,11 @@ public class SignInActivity extends AppCompatActivity implements ConnectionCallb
     public static final String EXTRA_JETPACK_MESSAGE_AUTH = "EXTRA_JETPACK_MESSAGE_AUTH";
     public static final String EXTRA_IS_AUTH_ERROR = "EXTRA_IS_AUTH_ERROR";
     public static final String EXTRA_PREFILL_URL = "EXTRA_PREFILL_URL";
+    public static final String MAGIC_LOGIN = "magic-login";
+    public static final String TOKEN_PARAMETER = "token";
 
     private SmartLockHelper mSmartLockHelper;
+    private ProgressDialog mProgressDialog;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -71,13 +84,68 @@ public class SignInActivity extends AppCompatActivity implements ConnectionCallb
     protected void onResume() {
         super.onResume();
         actionMode(getIntent().getExtras());
+        if (hasMagicLinkLoginIntent()) {
+            AnalyticsTracker.track(AnalyticsTracker.Stat.LOGIN_MAGIC_LINK_OPENED);
+            attemptLoginWithToken(getIntent().getData());
+        }
     }
 
-    protected void addSignInFragment() {
-        SignInFragment signInFragment = new SignInFragment();
-        FragmentTransaction fragmentTransaction = getSupportFragmentManager().beginTransaction();
-        fragmentTransaction.replace(R.id.fragment_container, signInFragment, SignInFragment.TAG);
-        fragmentTransaction.commit();
+    @Override
+    protected void onPause() {
+        super.onPause();
+        cancelProgressDialog();
+    }
+
+    @Override
+    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
+        super.onActivityResult(requestCode, resultCode, data);
+
+        if (requestCode == SHOW_CERT_DETAILS) {
+            getSignInFragment().askForSslTrust();
+        } else if (requestCode == SMART_LOCK_SAVE) {
+            if (resultCode == RESULT_OK) {
+                AnalyticsTracker.track(Stat.LOGIN_AUTOFILL_CREDENTIALS_UPDATED);
+                AppLog.d(T.NUX, "Credentials saved");
+            } else {
+                AppLog.d(T.NUX, "Credentials save cancelled");
+            }
+        } else if (requestCode == SMART_LOCK_READ) {
+            if (resultCode == RESULT_OK) {
+                AppLog.d(T.NUX, "Credentials retrieved");
+                Credential credential = data.getParcelableExtra(Credential.EXTRA_KEY);
+                getSignInFragment().onCredentialRetrieved(credential);
+            } else {
+                AppLog.e(T.NUX, "Credential read failed");
+            }
+        }
+    }
+
+    private void cancelProgressDialog() {
+        if (mProgressDialog != null && mProgressDialog.isShowing()) {
+            mProgressDialog.cancel();
+        }
+    }
+
+    private void attemptLoginWithToken(Uri uri) {
+        getSignInFragment().setToken(uri.getQueryParameter(TOKEN_PARAMETER));
+        MagicLinkSignInFragment magicLinkSignInFragment = getSignInFragment();
+        slideInFragment(magicLinkSignInFragment, false);
+
+        mProgressDialog = ProgressDialog
+                .show(this, "", getString(R.string.logging_in), true, true, new DialogInterface.OnCancelListener() {
+                    @Override
+                    public void onCancel(DialogInterface dialog) {
+                        getSignInFragment().setToken("");
+                    }
+                });
+        mProgressDialog.show();
+    }
+
+    private boolean hasMagicLinkLoginIntent() {
+        String action = getIntent().getAction();
+        Uri uri = getIntent().getData();
+
+        return Intent.ACTION_VIEW.equals(action) && uri != null && uri.getHost().contains(MAGIC_LOGIN);
     }
 
     protected void actionMode(Bundle extras) {
@@ -105,34 +173,11 @@ public class SignInActivity extends AppCompatActivity implements ConnectionCallb
         }
     }
 
-    @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-
-        if (requestCode == SHOW_CERT_DETAILS) {
-            getSignInFragment().askForSslTrust();
-        } else if (requestCode == SMART_LOCK_SAVE) {
-            if (resultCode == RESULT_OK) {
-                AnalyticsTracker.track(Stat.LOGIN_AUTOFILL_CREDENTIALS_UPDATED);
-                AppLog.d(T.NUX, "Credentials saved");
-            } else {
-                AppLog.d(T.NUX, "Credentials save cancelled");
-            }
-        } else if (requestCode == SMART_LOCK_READ) {
-            if (resultCode == RESULT_OK) {
-                AppLog.d(T.NUX, "Credentials retrieved");
-                Credential credential = data.getParcelableExtra(Credential.EXTRA_KEY);
-                getSignInFragment().onCredentialRetrieved(credential);
-            } else {
-                AppLog.e(T.NUX, "Credential read failed");
-            }
-        }
-    }
-
-    public SignInFragment getSignInFragment() {
-        SignInFragment signInFragment = (SignInFragment) getSupportFragmentManager().findFragmentByTag(SignInFragment.TAG);
+    public MagicLinkSignInFragment getSignInFragment() {
+        MagicLinkSignInFragment signInFragment =
+                (MagicLinkSignInFragment) getSupportFragmentManager().findFragmentByTag(MagicLinkSignInFragment.TAG);
         if (signInFragment == null) {
-            return new SignInFragment();
+            return new MagicLinkSignInFragment();
         } else {
             return signInFragment;
         }
@@ -140,6 +185,43 @@ public class SignInActivity extends AppCompatActivity implements ConnectionCallb
 
     public SmartLockHelper getSmartLockHelper() {
         return mSmartLockHelper;
+    }
+
+    private void saveEmailToAccount(String email) {
+        Account account = AccountHelper.getDefaultAccount();
+        account.setUserName(email);
+        account.save();
+    }
+
+    private void popBackStackToSignInFragment() {
+        FragmentManager fragmentManager = getSupportFragmentManager();
+        while (fragmentManager.getBackStackEntryCount() > 1) {
+            fragmentManager.popBackStackImmediate();
+        }
+
+        getSupportFragmentManager().popBackStack();
+    }
+
+    protected void addSignInFragment() {
+        MagicLinkSignInFragment signInFragment = new MagicLinkSignInFragment();
+        FragmentTransaction fragmentTransaction = getSupportFragmentManager().beginTransaction();
+        fragmentTransaction.replace(R.id.fragment_container, signInFragment, MagicLinkSignInFragment.TAG);
+        fragmentTransaction.commit();
+    }
+
+    private void slideInFragment(Fragment fragment) {
+        slideInFragment(fragment, true);
+    }
+
+    private void slideInFragment(Fragment fragment, boolean shouldAddToBackStack) {
+        FragmentTransaction fragmentTransaction = getSupportFragmentManager().beginTransaction();
+        fragmentTransaction.setCustomAnimations(R.anim.activity_slide_in_from_right, R.anim.activity_slide_out_to_left,
+                R.anim.activity_slide_in_from_left, R.anim.activity_slide_out_to_right);
+        fragmentTransaction.replace(R.id.fragment_container, fragment);
+        if (shouldAddToBackStack) {
+            fragmentTransaction.addToBackStack(null);
+        }
+        fragmentTransaction.commitAllowingStateLoss();
     }
 
     @Override
@@ -171,5 +253,27 @@ public class SignInActivity extends AppCompatActivity implements ConnectionCallb
     @Override
     public void onConnectionSuspended(int i) {
         AppLog.d(T.NUX, "Google API client connection suspended");
+    }
+
+    @Override
+    public void onMagicLinkSent() {
+        MagicLinkSentFragment magicLinkSentFragment = new MagicLinkSentFragment();
+        slideInFragment(magicLinkSentFragment);
+    }
+
+    @Override
+    public void onEnterPasswordRequested() {
+        AnalyticsTracker.track(AnalyticsTracker.Stat.LOGIN_MAGIC_LINK_EXITED);
+        getSignInFragment().setShouldShowPassword(true);
+
+        popBackStackToSignInFragment();
+    }
+
+    @Override
+    public void onMagicLinkRequestSuccess(String email) {
+        saveEmailToAccount(email);
+
+        MagicLinkRequestFragment magicLinkRequestFragment = MagicLinkRequestFragment.newInstance(email);
+        slideInFragment(magicLinkRequestFragment);
     }
 }
